Don't assume local and remote have the same name.
Kellan Elliot-McCrea pointed out an undocumented assumption that the last component of the local and remote pathnames are the same.  This is a bug not an undocumented assumption.  The fix involves the --relative option to rsync(1) and the placement of "./" at the end of the source path in each of the two preliminary rsync calls.